test: add tests for track trace hooks
