+++ tests/Pest.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+namespace MyParcelNL\WooCommerce\Tests;
+
 /**
  * Global Pest test configuration.
  *
@@ -13,5 +15,6 @@ use function MyParcelNL\Pdk\Tests\usesShared;
 
 require __DIR__ . '/../vendor/myparcelnl/pdk/tests/Pest.php';
 require __DIR__ . '/mock_class_map.php';
+require __DIR__ . '/functions.php';
 
 usesShared(new ClearContainerCache())->in(__DIR__);
